requirements.txt
fastapi
uvicorn
httpx
yfinance
sqlalchemy
apscheduler
beautifulsoup4
lxml
googlesearch-python
pydantic-settings
